+++ ui/admin/app/components/form/storage-bucket/aws/index.hbs
@@ -1,8 +1,6 @@
-{{! <h1>AWS</h1> }}
-
 <Rose::Form
   @onSubmit={{@submit}}
-  @cancel={{this.cancel}}
+  @cancel={{@cancel}}
   @disabled={{@model.isSaving}}
   @showEditToggle={{if @model.isNew false true}}
   as |form|
