+++ ui/admin/app/components/form/storage-bucket/index.hbs
@@ -12,6 +12,7 @@
   scopes=@scopes
   pluginTypes=this.pluginTypes
   changePluginType=@changePluginType
+  changeCredentialType=@changeCredentialType
   credentials=this.credentials
   selectedCredentialType=this.selectedCredentialType
   updateTypeSelection=this.updateTypeSelection
