+++ ui/admin/app/components/form/storage-bucket/minio/index.hbs
@@ -1,6 +1,6 @@
 <Rose::Form
   @onSubmit={{@submit}}
-  @cancel={{this.cancel}}
+  @cancel={{@cancel}}
   @disabled={{@model.isSaving}}
   @showEditToggle={{if @model.isNew false true}}
   as |form|
