+++ ui/admin/app/controllers/scopes/scope/storage-buckets/storage-bucket/index.js
@@ -1,0 +1,6 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeStorageBucketsStorageBucketIndexController extends Controller {
+  @controller('scopes/scope/storage-buckets/index')
+  storageBuckets;
+}
