+++ ui/admin/app/routes/scopes/scope/storage-buckets.js
@@ -5,10 +5,6 @@
 
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
-import { action } from '@ember/object';
-import { loading } from 'ember-loading';
-import { confirm } from 'core/decorators/confirm';
-import { notifySuccess, notifyError } from 'core/decorators/notify';
 
 export default class ScopesScopeStorageBucketsRoute extends Route {
   // =services
@@ -17,8 +13,6 @@ export default class ScopesScopeStorageBucketsRoute extends Route {
   @service session;
   @service can;
   @service router;
-
-  // =attributes
 
   // =methods
 
@@ -43,60 +37,5 @@ export default class ScopesScopeStorageBucketsRoute extends Route {
     ) {
       return this.store.query('storage-bucket', { scope_id, recursive: true });
     }
-  }
-
-  // =actions
-
-  /**
-   * Handle save
-   * @param {StorageBucketModel} storageBucket
-   */
-  @action
-  @loading
-  @notifyError(({ message }) => message)
-  @notifySuccess('notifications.save-success')
-  async save(storageBucket) {
-    await storageBucket.save();
-    await this.router.transitionTo(
-      'scopes.scope.storage-buckets.storage-bucket',
-      storageBucket,
-    );
-    this.refresh();
-  }
-
-  /**
-   * Rollback changes on storage buckets.
-   * @param {StorageBucketModel} storageBucket
-   */
-  @action
-  cancel(storageBucket) {
-    const { isNew } = storageBucket;
-    storageBucket.rollbackAttributes();
-    if (isNew) this.router.transitionTo('scopes.scope.storage-buckets');
-  }
-
-  /**
-   * Deletes the storage bucket.
-   * @param {StorageBucketModel} storageBucket
-   */
-  @action
-  @loading
-  @confirm('resources.storage-bucket.questions.delete-storage-bucket.message', {
-    title: 'resources.storage-bucket.questions.delete-storage-bucket.title',
-    confirm: 'resources.storage-bucket.actions.delete',
-  })
-  @notifyError(({ message }) => message, { catch: true })
-  @notifySuccess('notifications.delete-success')
-  async delete(storageBucket) {
-    await storageBucket.destroyRecord();
-  }
-
-  /**
-   * Updates type of credential
-   * @param {string} credentialType
-   */
-  @action
-  changeType(storageBucket, credentialType) {
-    storageBucket.credentialType = credentialType;
   }
 }
