+++ ui/admin/app/templates/scopes/scope/storage-buckets/index.hbs
@@ -113,7 +113,7 @@
                       <dd.Interactive
                         @text={{t 'resources.storage-bucket.actions.delete'}}
                         @color='critical'
-                        {{on 'click' (route-action 'delete' B.data)}}
+                        {{on 'click' (fn this.delete B.data)}}
                       />
                     {{/if}}
                   </Hds::Dropdown>
