+++ ui/admin/app/templates/scopes/scope/storage-buckets/new.hbs
@@ -24,10 +24,13 @@
   <page.body>
     <Form::StorageBucket
       @model={{@model}}
-      @submit={{route-action 'save' @model}}
-      @cancel={{route-action 'cancel' @model}}
+      @submit={{fn this.storageBuckets.save @model}}
+      @cancel={{fn this.storageBuckets.cancel @model}}
       @scopes={{this.scopes}}
-      @changeType={{route-action 'changeType' @model}}
+      @changeCredentialType={{fn
+        this.storageBuckets.changeCredentialType
+        @model
+      }}
       @changePluginType={{this.changePluginType}}
     />
   </page.body>
