+++ ui/admin/app/templates/scopes/scope/storage-buckets/storage-bucket/index.hbs
@@ -27,10 +27,13 @@
   <page.body>
     <Form::StorageBucket
       @model={{@model}}
-      @submit={{route-action 'save' @model}}
-      @cancel={{route-action 'cancel' @model}}
+      @submit={{fn this.storageBuckets.save @model}}
+      @cancel={{fn this.storageBuckets.cancel @model}}
       @scopes={{this.scopes}}
-      @changeType={{route-action 'changeType' @model}}
+      @changeCredentialType={{fn
+        this.storageBuckets.changeCredentialType
+        @model
+      }}
     />
   </page.body>
 </Rose::Layout::Page>
