+++ ui/admin/tests/unit/controllers/scopes/scope/storage-buckets/storage-bucket/index-test.js
@@ -1,0 +1,17 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'admin/tests/helpers';
+
+module(
+  'Unit | Controller | scopes/scope/storage-buckets/storage-bucket/index',
+  function (hooks) {
+    setupTest(hooks);
+
+    // TODO: Replace this with your real tests.
+    test('it exists', function (assert) {
+      let controller = this.owner.lookup(
+        'controller:scopes/scope/storage-buckets/storage-bucket/index',
+      );
+      assert.ok(controller);
+    });
+  },
+);
